refactor: create nextjs preset utils
