+++ lib/presets/custom/next/deliver/prebuild.js
@@ -1,4 +1,7 @@
+import { readdir, stat, mkdir, rm, copyFile } from 'fs/promises';
+import { join, extname, basename, dirname } from 'path';
 import { gte, coerce, valid } from 'semver';
+
 import {
   exec,
   feedback,
@@ -7,18 +10,7 @@ import {
   copyDirectory,
   Manifest,
 } from '#utils';
-
-import {
-  readdir,
-  stat,
-  mkdir,
-  readFile,
-  rm,
-  copyFile,
-  access,
-  constants,
-} from 'fs/promises';
-import { join, extname, basename, dirname, resolve } from 'path';
+import { getNextConfig, readManifestFile } from '../utils.next.js';
 
 const packageManager = await getPackageManager();
 
@@ -91,18 +83,6 @@ async function moveFiles(directory) {
       }
     }),
   );
-}
-
-/**
- * Get a project manifest generated after build
- * @param {string} file - file name
- * @returns {object} - the manifest
- */
-async function readManifestFile(file) {
-  const manifestPath = resolve(process.cwd(), '.next', file);
-  const manifest = await readFile(manifestPath, 'utf-8');
-
-  return JSON.parse(manifest);
 }
 
 /**
@@ -190,54 +170,6 @@ function isANewerVersion(version) {
 }
 
 /**
- * Check if a file exists
- * @param {string} path - file path
- * @returns {boolean} indicates if exists or not
- */
-async function fileExists(path) {
-  try {
-    await access(path, constants.F_OK);
-  } catch (err) {
-    // file does not exists
-    if (err.code === 'ENOENT') {
-      return false;
-    }
-    // other error cases
-    feedback.prebuild.error('Error reading file:', err);
-    throw err;
-  }
-
-  return true;
-}
-
-/**
- * Get next config file
- * @returns {object} - next config as a JSON
- */
-async function getNextConfig() {
-  try {
-    let configFile = null;
-    const jsConfigExists = await fileExists('next.config.js');
-    const mjsConfigExists = await fileExists('next.config.mjs');
-    if (jsConfigExists) {
-      configFile = 'next.config.js';
-    } else if (mjsConfigExists) {
-      configFile = 'next.config.mjs';
-    } else {
-      feedback.prebuild.info('Nextjs config file does not exists!');
-      return null;
-    }
-
-    const configPath = join(process.cwd(), configFile);
-    const configModule = await import(configPath);
-
-    return configModule.default;
-  } catch (error) {
-    throw Error('Error reading next config file:', error);
-  }
-}
-
-/**
  * Validates if static site mode is enabled in next config.
  * @param {object} nextConfig - the config as JSON object.
  */
@@ -264,8 +196,6 @@ async function prebuild() {
   // new build format
   if (isANewerVersion(nextVersion)) {
     const nextConfig = await getNextConfig();
-    console.log('*** Config:');
-    console.log(nextConfig);
 
     validateStaticSiteMode(nextConfig);
